requirements.txt
openai
azure-identity
tiktoken
numpy
scikit-learn
fastapi
uvicorn
streamlit
python-dotenv